Adding references and notes
I've updated the references.bib file to include the references for the Adam paper and the earlier version of the Perception paper. I've added references to the Perception paper at the bottom of the tex files 6,7,8,9, and also I've added a reference to the Adam paper in tex file 13. I made a note of the limitations of the Perceptron in tex file 7. I made a note about learning rate for gradient descent in tex file 9 as well.